Create abstract base tests
