+++ tests/Functional/AbstractBaseWebTestCase.php
@@ -1,0 +1,19 @@
+<?php
+
+declare(strict_types=1);
+
+namespace App\Tests\Functional;
+
+use Symfony\Bundle\FrameworkBundle\KernelBrowser;
+
+abstract class AbstractBaseWebTestCase extends AbstractBaseFunctionalTest
+{
+    protected KernelBrowser $client;
+
+    protected function setUp(): void
+    {
+        parent::setUp();
+
+        $this->client = static::createClient();
+    }
+}
